refactor: 코드 리뷰 반영 #103
